[wip] Create new export/collecting zone infrastructure
The new *2 defines in zone.pp use export/collect to send the
necessary infrastructure and the contents of the zone to the
bind::master.

This is still Work In Progress, since I need to completely re-do
all my zones to use this infrastructure. On the up-side, this will
allow me to put the DNS records where they origin.